isolated rootsystem and world
